WIP: jpms module-info added
