+++ native/pom.xml
@@ -1,46 +1,55 @@
 <project>
-    <modelVersion>4.0.0</modelVersion>
-    <groupId>org.jetbrains.skija</groupId>
-    <artifactId>skija-native</artifactId>
-    <version>0.0.0-SNAPSHOT</version>
-    <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-    <dependencies>
-        <dependency>
-            <groupId>org.jetbrains.skija</groupId>
-            <artifactId>skija-shared</artifactId>
-            <version>0.0.0-SNAPSHOT</version>
-        </dependency>
-    </dependencies>
-    <build>
-        <sourceDirectory>non-existent</sourceDirectory>
-        <resources>
-            <resource>
-                <directory>build</directory>
-                <includes>
-                    <include>libskija_x64.dylib</include>
-                    <include>libskija_arm64.dylib</include>
-                    <include>libskija.so</include>
-                    <include>skija.dll</include>
-                    <include>icudtl.dat</include>
-                    <include>skija.version</include>
-                </includes>
-            </resource>
-        </resources>
-    </build>
-    <repositories>
-        <repository>
-            <id>space-maven</id>
-            <url>https://packages.jetbrains.team/maven/p/skija/maven</url>
-        </repository>
-    </repositories>
-    <distributionManagement>
-        <repository>
-            <id>space-maven</id>
-            <url>https://packages.jetbrains.team/maven/p/skija/maven</url>
-        </repository>
-    </distributionManagement>
+	<modelVersion>4.0.0</modelVersion>
+	<groupId>org.jetbrains.skija</groupId>
+	<artifactId>skija-native</artifactId>
+	<version>0.0.0-local</version>
+	<properties>
+		<maven.compiler.source>11</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+	</properties>
+	<dependencies>
+		<dependency>
+			<groupId>org.jetbrains.skija</groupId>
+			<artifactId>skija-shared</artifactId>
+			<version>0.0.0-local</version>
+		</dependency>
+	</dependencies>
+	<build>
+		<sourceDirectory>src/main/java</sourceDirectory>
+		<resources>
+			<resource>
+				<!-- internal windows testing only -->
+				<directory>src/main/resources</directory>
+				<includes>
+					<include>skija.dll</include>
+					<include>icudtl.dat</include>
+					<include>skija.version</include>
+				</includes>
+			</resource>
+			<resource>
+				<directory>build</directory>
+				<includes>
+					<include>libskija_x64.dylib</include>
+					<include>libskija_arm64.dylib</include>
+					<include>libskija.so</include>
+					<include>skija.dll</include>
+					<include>icudtl.dat</include>
+					<include>skija.version</include>
+				</includes>
+			</resource>
+		</resources>
+	</build>
+	<repositories>
+		<repository>
+			<id>space-maven</id>
+			<url>https://packages.jetbrains.team/maven/p/skija/maven</url>
+		</repository>
+	</repositories>
+	<distributionManagement>
+		<repository>
+			<id>space-maven</id>
+			<url>https://packages.jetbrains.team/maven/p/skija/maven</url>
+		</repository>
+	</distributionManagement>
 </project>
