+++ native/script/package.py
@@ -19,6 +19,9 @@ def main():
   with open('pom.xml', 'w') as f:
     f.write(pom.replace('0.0.0-SNAPSHOT', rev).replace('skija-native', common.skija_native_artifact_id))
 
+  with open(os.path.join('src', 'main', 'java', 'module-info.java'), 'w') as f:
+    f.write(pom.replace('skija.windows', common.skija_platform_module_id))
+
   mvn = 'mvn.cmd' if common.system == 'windows' else 'mvn'
   common.check_call([
     mvn,
