+++ native/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module skija.windows {
+  
+  requires transitive skija.shared;
+  
+  opens org.jetbrains.skija.platform to skija.shared;
+
+}
