+++ native/src/main/java/org/jetbrains/skija/platform/LibraryLinux.java
@@ -1,0 +1,7 @@
+package org.jetbrains.skija.platform;
+
+public class LibraryLinux {
+
+  private LibraryLinux() {}
+
+}
