+++ native/src/main/java/org/jetbrains/skija/platform/LibraryMacosArm64.java
@@ -1,0 +1,7 @@
+package org.jetbrains.skija.platform;
+
+public class LibraryMacosArm64 {
+
+  private LibraryMacosArm64() {}
+
+}
