+++ native/src/main/java/org/jetbrains/skija/platform/LibraryMacosX64.java
@@ -1,0 +1,7 @@
+package org.jetbrains.skija.platform;
+
+public class LibraryMacosX64 {
+
+  private LibraryMacosX64() {}
+
+}
