+++ native/src/test/java/org/jetbrains/skija/platform/TestTest.java
@@ -1,0 +1,47 @@
+package org.jetbrains.skija.platform;
+
+import static org.jetbrains.skija.platform.test.TestRunner.assertArrayEquals;
+import static org.jetbrains.skija.platform.test.TestRunner.assertEquals;
+import static org.jetbrains.skija.platform.test.TestRunner.assertNotEquals;
+import static org.jetbrains.skija.platform.test.TestRunner.assertThrows;
+import static org.jetbrains.skija.platform.test.TestRunner.popStack;
+import static org.jetbrains.skija.platform.test.TestRunner.pushStack;
+
+import java.util.NoSuchElementException;
+
+import org.jetbrains.skija.platform.test.*;
+
+public class TestTest implements Executable {
+    @Override
+    public void execute() throws Exception {
+        TestRunner.testMethod(this, "abc");
+        throw new Exception("Unexpected in execute");
+    }
+
+    public void abc() throws Exception {
+        assertEquals(true, true);
+        assertEquals(true, false);
+
+        assertNotEquals(null, false);
+        assertNotEquals(true, true);
+
+        assertArrayEquals(new Integer[] {1, 2, 3},
+                          new Integer[] {1, 2, 3});
+        assertArrayEquals(new Integer[] {1, 2, 3},
+                          new Integer[] {1, 2});
+        assertArrayEquals(new Integer[] {1, 2, 3},
+                          new Integer[] {1, 2, 4});
+
+        assertThrows(NoSuchElementException.class, ()-> { throw new NoSuchElementException(); });
+        assertThrows(NoSuchElementException.class, ()-> { throw new Exception("abc"); });
+        assertThrows(NoSuchElementException.class, ()-> { System.gc(); });
+
+        for (int i = 0; i < 10; ++i) {
+            pushStack("i = " + i);
+            assertEquals(0, i % 2);
+            popStack();
+        }
+
+        throw new Exception("Unexpected in method");
+    }
+}
