+++ native/src/test/java/org/jetbrains/skija/platform/test/Executable.java
@@ -1,0 +1,6 @@
+package org.jetbrains.skija.platform.test;
+
+@FunctionalInterface
+public interface Executable {
+    void execute() throws Exception;
+}
