--- shared/src/test/java/org/jetbrains/skija/test/Executable.java
@@ -1,6 +1,0 @@
-package org.jetbrains.skija.test;
-
-@FunctionalInterface
-public interface Executable {
-    void execute() throws Exception;
-}
